restructure `nn` files
