+++ nequip/nn/pair_potential.py
@@ -6,8 +6,8 @@ from e3nn.util.jit import compile_mode
 
 from nequip.data import AtomicDataDict
 from nequip.data.misc import chemical_symbols_to_atomic_numbers_dict
-from ._util import scatter
 from ._graph_mixin import GraphModuleMixin
+from .utils import scatter
 from nequip.utils import conditional_torchscript_jit
